requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
empyrical==0.5.5
executing==1.2.0
fonttools==4.41.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
importlib-resources==6.0.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
multitasking==0.0.11
numpy==1.25.1
packaging==23.1
pandas==1.2.4
pandas-datareader==0.10.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
QuantStats==0.0.62
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
webencodings==0.5.1
yfinance==0.2.24
zipp==3.16.2
